third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/setting-allowpaymentrequest.https.sub-expected.txt

This is a testharness.js-based test.
FAIL PaymentRequest setting allowpaymentrequest after load and then navigating assert_array_equals: before navigation property 0, expected true but got false
Harness: the test ran to completion.
